moved stuff around, added 3d benchmark

+++ src/main/java/com/github/hageldave/fftbenchmark/interfaces/FFT3D.java
@@ -2,8 +2,8 @@ package com.github.hageldave.fftbenchmark.interfaces;
 
 public interface FFT3D {
 	
-	public void doubleArrayFFT_SetDC2Zero_3D(double[] array, double[] result, int width, int height, int depth);
+	public void doubleArrayFFT_SetDC2Zero_3D(float[] array, float[] result, int width, int height, int depth);
 	
-	public void double3DArrayFFT_SetDC2Zero_3D(double[][][] array, double[][][] result, int width, int height, int depth);
+	public void double3DArrayFFT_SetDC2Zero_3D(float[][][] array, float[][][] result, int width, int height, int depth);
 
 }
